requirements.txt
PyYAML==6.0.1
Jinja2==3.1.4
qrcode[pil]==7.4.2
Pillow==10.4.0 